add simpletest, remove buttontest
